requirements.txt
fastapi
uvicorn[standard]
pydantic
python-dotenv
firebase-admin
sendgrid
google-generativeai>=0.3.1
python-pushover